% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.R
\name{download_swf}
\alias{download_swf}
\title{Download SWF Spreadsheet from CS-DocDB.}
\usage{
download_swf(ver, username, passwd)
}
\arguments{
\item{ver}{The version of CS-doc-3799}

\item{username}{The username you use for cd-docdb}

\item{passwd}{The password you use for cd-docdb}
}
\value{
The name of the temporary file
}
\description{
Download the SWF spreadsheet from CS-doc-3799
\url{http://cd-docdb.fnal.gov:8080/cgi-bin/ShowDocument?docid=3799}. You must
supply the version number you want (no default) and the access username and
password for DocDB. A temporary file is created and that name is returned.
}

