% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf.R
\name{swf_df}
\alias{swf_df}
\title{Extract SWF ata frame from SWF spreadsheet.}
\usage{
swf_df(excelFile)
}
\arguments{
\item{swfFile}{The name of the file with the SWF spreadsheet}
}
\value{
A data frame
}
\description{
Extract the SWF data frame from the SWF and fix the columns appropriately.
}
\examples{
\dontrun{
download_swf(422, "??", "??") \%>\% swf_df -> swfdf
}
}

